refactor: SelectJdbcTemplate를 JdbcTemplate으로 통합
