+++ src/main/java/next/dao/JdbcTemplate.java
@@ -4,7 +4,10 @@ import core.jdbc.ConnectionManager;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 
 public abstract class JdbcTemplate {
     public void update(String sql) throws SQLException {
@@ -14,7 +17,6 @@ public abstract class JdbcTemplate {
             con = ConnectionManager.getConnection();
             pstmt = con.prepareStatement(sql);
             setValues(pstmt);
-
             pstmt.executeUpdate();
         } finally {
             if (pstmt != null) {
@@ -26,6 +28,64 @@ public abstract class JdbcTemplate {
             }
         }
     }
+    public List<Object> query(String sql) throws SQLException {
+        Connection con = null;
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+        try {
+            con = ConnectionManager.getConnection();
+            pstmt = con.prepareStatement(sql);
+            setValues(pstmt);
+            rs = pstmt.executeQuery();
+
+            List<Object> list = new ArrayList<>();
+            while (rs.next()) {
+                list.add(mapRow(rs));
+            }
+            return list;
+        } finally {
+            if (rs != null) {
+                rs.close();
+            }
+            if (pstmt != null) {
+                pstmt.close();
+            }
+            if (con != null) {
+                con.close();
+            }
+        }
+    }
+
+    public Object queryForObject(String sql) throws SQLException {
+        Connection con = null;
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+        try {
+            con = ConnectionManager.getConnection();
+            pstmt = con.prepareStatement(sql);
+            setValues(pstmt);
+            rs = pstmt.executeQuery();
+
+            Object o = null;
+            if (rs.next()) {
+                o = mapRow(rs);
+            }
+
+            return o;
+        } finally {
+            if (rs != null) {
+                rs.close();
+            }
+            if (pstmt != null) {
+                pstmt.close();
+            }
+            if (con != null) {
+                con.close();
+            }
+        }
+    }
 
     protected abstract void setValues(PreparedStatement pstmt) throws SQLException;
+
+    protected abstract Object mapRow(ResultSet rs) throws SQLException;
 }
