--- src/main/java/next/dao/SelectJdbcTemplate.java
@@ -1,73 +1,0 @@
-package next.dao;
-
-import core.jdbc.ConnectionManager;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-
-public abstract class SelectJdbcTemplate {
-    public List<Object> query(String sql) throws SQLException {
-        Connection con = null;
-        PreparedStatement pstmt = null;
-        ResultSet rs = null;
-        try {
-            con = ConnectionManager.getConnection();
-            pstmt = con.prepareStatement(sql);
-            setValues(pstmt);
-            rs = pstmt.executeQuery();
-
-            List<Object> list = new ArrayList<>();
-            while (rs.next()) {
-                list.add(mapRow(rs));
-            }
-            return list;
-        } finally {
-            if (rs != null) {
-                rs.close();
-            }
-            if (pstmt != null) {
-                pstmt.close();
-            }
-            if (con != null) {
-                con.close();
-            }
-        }
-    }
-
-    public Object queryForObject(String sql) throws SQLException {
-        Connection con = null;
-        PreparedStatement pstmt = null;
-        ResultSet rs = null;
-        try {
-            con = ConnectionManager.getConnection();
-            pstmt = con.prepareStatement(sql);
-            setValues(pstmt);
-            rs = pstmt.executeQuery();
-
-            Object o = null;
-            if (rs.next()) {
-                o = mapRow(rs);
-            }
-
-            return o;
-        } finally {
-            if (rs != null) {
-                rs.close();
-            }
-            if (pstmt != null) {
-                pstmt.close();
-            }
-            if (con != null) {
-                con.close();
-            }
-        }
-    }
-
-    protected abstract void setValues(PreparedStatement pstmt) throws SQLException;
-
-    protected abstract Object mapRow(ResultSet rs) throws SQLException;
-}
